# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Service for Codex",
        "title": "API Service",
        "contact": {}
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
        "/admin/attributes/": {
            "post": {
                "description": "Adds Attribute Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attributes"
                ],
                "summary": "Add Attribute",
                "parameters": [
                    {
                        "description": "New Attribute",
                        "name": "newAttribute",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddAttributeDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates Attribute Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attributes"
                ],
                "summary": "Update Attribute",
                "parameters": [
                    {
                        "description": "Update Attribute",
                        "name": "updateAttribute",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAttributeDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/admin/attributes/:id": {
            "delete": {
                "description": "Delete Attributes from DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attributes"
                ],
                "summary": "Delete Attribute",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attribute ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/rewards/": {
            "post": {
                "description": "Adds Reward Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reward"
                ],
                "summary": "Add Reward",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Reward Image File",
                        "name": "imageFile",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "Description",
                        "in": "formData"
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "rewardType",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "sellerFee",
                        "in": "formData"
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "symbol",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates Reward Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reward"
                ],
                "summary": "Update Reward",
                "parameters": [
                    {
                        "type": "string",
                        "name": "Description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "rewardType",
                        "in": "formData"
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "symbol",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Course Image File",
                        "name": "imageFile",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/admin/rewards/{id}": {
            "delete": {
                "description": "Delete Rewards from DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reward"
                ],
                "summary": "Delete Reward",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reward ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/private/attributes/": {
            "get": {
                "description": "Retrieves all attribute based on the provided query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attributes"
                ],
                "summary": "Get All Attributes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attribute ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reward ID",
                        "name": "rewardID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "TraitType of Attribute",
                        "name": "traitType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/private/rewards/": {
            "get": {
                "description": "Retrieves all rewards based on the provided query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reward"
                ],
                "summary": "Get All Rewards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reward ID",
                        "name": "rewardID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reward Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reward Symbol",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reward Type",
                        "name": "rewardType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/private/rewards/{id}": {
            "get": {
                "description": "Retrieves one reward.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reward"
                ],
                "summary": "Get Reward By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reward ID",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Reward Attribute Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reward Attribute Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddAttributeDTO": {
            "type": "object",
            "required": [
                "rewardID",
                "traitType",
                "value"
            ],
            "properties": {
                "rewardID": {
                    "type": "string"
                },
                "traitType": {
                    "type": "string",
                    "maxLength": 30
                },
                "value": {
                    "type": "string",
                    "maxLength": 30
                }
            }
        },
...
        "dto.UpdateAttributeDTO": {
            "type": "object",
            "required": [
                "rewardID"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "rewardID": {
                    "type": "string"
                },
                "traitType": {
                    "type": "string",
                    "maxLength": 30
                },
                "value": {
                    "type": "string",
                    "maxLength": 30
                }
            }
        },
...
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "dataCount": {
                    "type": "integer"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}